+++ app.js
@@ -76,6 +76,28 @@ app.get('/', (req, res) => {
   res.redirect('/service-manual');
 });
 
+// The practices pages have moved or been deleted
+// Temporary redirects incase anyone still visits /practices pages
+app.get('/service-manual/practices/create-content-for-users-with-low-health-literacy', (req, res) => {
+  res.redirect('/service-manual/content/create-content-for-users-with-low-health-literacy');
+});
+
+app.get('/service-manual/practices/create-content-for-users-with-low-health-literacy/use-a-readability-tool-to-prioritise-content', (req, res) => {
+  res.redirect('/service-manual/content/create-content-for-users-with-low-health-literacy/use-a-readability-tool-to-prioritise-content');
+});
+
+app.get('/service-manual/practices', (req, res) => {
+  res.redirect('/service-manual');
+});
+
+app.get('/service-manual/practices/make-your-service-accessible', (req, res) => {
+  res.redirect('/service-manual/accessibility');
+});
+
+app.get('/service-manual/content/writing-for-accessibility', (req, res) => {
+  res.redirect('/service-manual/accessibility/content');
+});
+
 // Automatically route pages
 app.get(/^([^.]+)$/, (req, res, next) => {
   routing.matchRoutes(req, res, next);